separate VideoGalleryResponsiveHorizontalGallery component
